Update to v69.420.004
Hello to all,

This update fixes some bugs found:
-Admin Login Redirect fix
- User Login Redirect Fix
- Wen runing Cron task to check payments in Doge, is checks now every unxpendable betwin 0 Doge and Above

Optimizations:
- Added CoinGecko Doge/Fiat to show current price
- Wen adding product to cart, auto display cart.

+++ dogegarden.sql
@@ -7,15 +7,15 @@ SET sql_mode = 'NO_AUTO_VALUE_ON_ZERO';
 
 DROP TABLE IF EXISTS `banners`;
 CREATE TABLE `banners` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint(20) NOT NULL AUTO_INCREMENT,
   `lang` varchar(2) NOT NULL DEFAULT 'en',
-  `id_cat` int(11) DEFAULT NULL,
-  `id_prod` int(11) DEFAULT NULL,
-  `id_page` int(11) DEFAULT NULL,
+  `id_cat` bigint(20) DEFAULT NULL,
+  `id_prod` bigint(20) DEFAULT NULL,
+  `id_page` bigint(20) DEFAULT NULL,
   `img` varchar(255) DEFAULT NULL,
   `video` varchar(255) DEFAULT NULL,
   `link` varchar(255) DEFAULT NULL,
-  `ord` int(11) NOT NULL,
+  `ord` bigint(20) NOT NULL,
   `active` int(1) DEFAULT '1',
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
@@ -23,11 +23,11 @@ CREATE TABLE `banners` (
 
 DROP TABLE IF EXISTS `cart`;
 CREATE TABLE `cart` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `id_shibe` int(11) NOT NULL,
+  `id` bigint(20) NOT NULL AUTO_INCREMENT,
+  `id_shibe` bigint(20) NOT NULL,
   `ip` varchar(255) DEFAULT NULL,
   `session` varchar(255) DEFAULT NULL,
-  `product_id` int(11) DEFAULT NULL,
+  `product_id` bigint(20) DEFAULT NULL,
   `qty` int(11) DEFAULT NULL,
   `date` datetime DEFAULT NULL,
   PRIMARY KEY (`id`)
@@ -36,10 +36,10 @@ CREATE TABLE `cart` (
 
 DROP TABLE IF EXISTS `categories`;
 CREATE TABLE `categories` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint(20) NOT NULL AUTO_INCREMENT,
   `lang` varchar(2) DEFAULT 'en',
   `icon` varchar(50) DEFAULT 'ellipsis-v',
-  `id_cat` int(11) DEFAULT '0',
+  `id_cat` bigint(20) DEFAULT '0',
   `title` varchar(255) DEFAULT NULL,
   `text` text,
   `img` varchar(255) DEFAULT NULL,
@@ -51,15 +51,15 @@ CREATE TABLE `categories` (
 
 DROP TABLE IF EXISTS `orders`;
 CREATE TABLE `orders` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `id_shibe` int(11) NOT NULL,
+  `id` bigint(20) NOT NULL AUTO_INCREMENT,
+  `id_shibe` bigint(20) NOT NULL,
   `doge_in_address` varchar(255) DEFAULT NULL,
   `doge_out_address` varchar(255) DEFAULT NULL,
   `tax` decimal(10,2) DEFAULT NULL,
   `total_doge` decimal(20,8) DEFAULT NULL,
   `doge_transaction_id` text,
-  `confirmations` int(11) NOT NULL DEFAULT '0',
-  `shipping` decimal(10,2) DEFAULT NULL,
+  `confirmations` bigint(20) NOT NULL DEFAULT '0',
+  `shipping` decimal(20,8) DEFAULT NULL,
   `products_json` longtext,
   `status` int(1) NOT NULL,
   `email_sent` int(1) NOT NULL,
@@ -70,13 +70,13 @@ CREATE TABLE `orders` (
 
 DROP TABLE IF EXISTS `pages`;
 CREATE TABLE `pages` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint(20) NOT NULL AUTO_INCREMENT,
   `lang` varchar(2) NOT NULL DEFAULT 'en',
-  `id_page` int(11) NOT NULL DEFAULT '0',
+  `id_page` bigint(20) NOT NULL DEFAULT '0',
   `type` int(10) NOT NULL DEFAULT '0',
   `title` varchar(255) DEFAULT NULL,
   `text` text,
-  `ord` int(11) DEFAULT '0',
+  `ord` bigint(20) DEFAULT '0',
   `active` int(1) DEFAULT '1',
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
@@ -84,19 +84,19 @@ CREATE TABLE `pages` (
 
 DROP TABLE IF EXISTS `products`;
 CREATE TABLE `products` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `id_cat` int(11) NOT NULL,
+  `id` bigint(20) NOT NULL AUTO_INCREMENT,
+  `id_cat` bigint(20) NOT NULL,
   `tax` decimal(10,2) NOT NULL,
   `doge` decimal(20,8) NOT NULL,
   `moon_new` decimal(10,2) DEFAULT '0.00',
   `moon_full` decimal(10,2) DEFAULT NULL,
-  `qty` int(11) DEFAULT '0',
+  `qty` bigint(11) DEFAULT '0',
   `weight` decimal(10,2) DEFAULT '0.00',
   `title` varchar(255) DEFAULT NULL,
   `text` text,
   `imgs` longtext,
   `highlighted` tinyint(1) DEFAULT '0',
-  `ord` int(11) DEFAULT '0',
+  `ord` bigint(20) DEFAULT '0',
   `active` int(1) NOT NULL DEFAULT '1',
   `date` datetime DEFAULT NULL,
   PRIMARY KEY (`id`)
@@ -105,7 +105,7 @@ CREATE TABLE `products` (
 
 DROP TABLE IF EXISTS `shibes`;
 CREATE TABLE `shibes` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint(20) NOT NULL AUTO_INCREMENT,
   `name` varchar(255) NOT NULL,
   `email` varchar(255) NOT NULL,
   `password` varchar(255) NOT NULL,
@@ -124,7 +124,7 @@ CREATE TABLE `shibes` (
 
 DROP TABLE IF EXISTS `shipping`;
 CREATE TABLE `shipping` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint(20) NOT NULL AUTO_INCREMENT,
   `country` varchar(2) NOT NULL DEFAULT 'en',
   `title` varchar(255) DEFAULT NULL,
   `text` varchar(255) DEFAULT NULL,
